Added exclusion for /home/*/.gvfs
We experienced a problem with rsync where nothing on the remote side was being deleted. The cause was this .gvfs directory. It's a FUSE mount, something new for Gnome. And its perms are very strange such that nobody can read it. This was causing rsync to doubt the file list and disable dest deletion.
The opinion out there is this mount dir isn't important to back up.